# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GalaxyBots.ai API - AI-powered corporate bot platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bots/{id}:
    get:
      operationId: getBot
      tags: [bots]
      summary: Get a bot by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
        "404":
          description: Bot not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
    Bot:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        department:
          type: string
        category:
          type: string
        description:
          type: string
        responsibilities:
          type: array
          items:
            type: string
        personality:
          type: string
        avatar:
          type: ["string", "null"]
        addonType:
          type: ["string", "null"]
        isAvailable:
          type: boolean
        isAiGenerated:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, name, title, department, category, description, responsibilities, personality, isAvailable, isAiGenerated, createdAt]
...