# lib/api-zod/src/generated/types/bot.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GalaxyBots.ai API - AI-powered corporate bot platform
 * OpenAPI spec version: 0.1.0
 */

export interface Bot {
  id: number;
  name: string;
  title: string;
  department: string;
  category: string;
  description: string;
  responsibilities: string[];
  personality: string;
  /** @nullable */
  avatar?: string | null;
  /** @nullable */
  addonType?: string | null;
  isAvailable: boolean;
  isAiGenerated: boolean;
  createdAt: Date;
}